Adding CSS
I attempted to add CSS for a center lined scrolling webpage, but I am
having issues because of the columons in the HTML. For now I am going to
push this so we can discuss if we like this approach.

I did not touch the poem page because I did not check permission ahead
of time to change what was already there.